+++ common/src/main/java/org/orecruncher/dsurround/mixins/core/MixinMusic.java
@@ -6,6 +6,7 @@ import org.orecruncher.dsurround.mixinutils.MixinHelpers;
 import org.spongepowered.asm.mixin.Final;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
+import org.spongepowered.asm.mixin.Unique;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
@@ -23,19 +24,21 @@ public class MixinMusic {
 
     @Inject(method = "getMinDelay()I", at = @At("HEAD"), cancellable = true)
     public void dsurround$getMinDelay(CallbackInfoReturnable<Integer> cir) {
-        if (MixinHelpers.musicOptions.reduceWaitTime != 0) {
-            var keepAmount = 100 - MixinHelpers.musicOptions.reduceWaitTime;
-            var newWaitTime = Mth.clamp((this.minDelay * keepAmount) / 100, 1, this.minDelay);
-            cir.setReturnValue(newWaitTime);
+        if (MixinHelpers.musicOptions.reduceWaitTime > 0) {
+            cir.setReturnValue(dsurround$calculateNewDelayThreshold(this.minDelay));
         }
     }
 
     @Inject(method = "getMaxDelay()I", at = @At("HEAD"), cancellable = true)
     public void dsurround$getMaxDelay(CallbackInfoReturnable<Integer> cir) {
-        if (MixinHelpers.musicOptions.reduceWaitTime != 0) {
-            var keepAmount = 100 - MixinHelpers.musicOptions.reduceWaitTime;
-            var newWaitTime = Mth.clamp((this.maxDelay * keepAmount) / 100, 1, this.maxDelay);
-            cir.setReturnValue(newWaitTime);
+        if (MixinHelpers.musicOptions.reduceWaitTime > 0) {
+            cir.setReturnValue(dsurround$calculateNewDelayThreshold(this.maxDelay));
         }
+    }
+
+    @Unique
+    private static int dsurround$calculateNewDelayThreshold(int currentThreshold) {
+        var keepAmount = 100 - MixinHelpers.musicOptions.reduceWaitTime;
+        return Mth.clamp((currentThreshold * keepAmount) / 100, 1, currentThreshold);
     }
 }