+++ fabric/src/main/java/org/orecruncher/fabric/FabricMod.java
@@ -1,23 +1,17 @@
 package org.orecruncher.fabric;
 
-import net.fabricmc.api.ClientModInitializer;
 import org.orecruncher.dsurround.Client;
 
 /**
  * Implements the Fabric specific binding to initialize the mod
  */
-public final class FabricMod implements ClientModInitializer {
+public final class FabricMod {
 
-    private final Client client;
-
-    public FabricMod() {
-        this.client = new Client();
-        this.client.construct();
+    public static void initialize() {
+        Client.initialize();
     }
 
-    @Override
-    public void onInitializeClient() {
-        // Boot the mod
-        this.client.initializeClient();
+    public static void initializeClient() {
+        Client.initializeClient();
     }
 }