+++ neoforge/src/main/java/org/orecruncher/neoforge/NeoForgeMod.java
@@ -7,7 +7,6 @@ import net.neoforged.fml.ModContainer;
 import net.neoforged.fml.ModList;
 import net.neoforged.fml.common.Mod;
 
-import net.neoforged.fml.event.lifecycle.FMLClientSetupEvent;
 import net.neoforged.neoforge.client.event.RegisterGuiLayersEvent;
 import net.neoforged.neoforge.client.gui.IConfigScreenFactory;
 import org.orecruncher.dsurround.Client;
@@ -18,15 +17,11 @@ import org.orecruncher.dsurround.lib.di.ContainerManager;
 @Mod(value = Constants.MOD_ID, dist = Dist.CLIENT)
 public final class NeoForgeMod {
 
-    private final Client client;
-
     public NeoForgeMod(ModContainer container, IEventBus modBus) {
-        modBus.addListener(this::onInitializeClient);
         modBus.addListener(this::onRegisterGuiLayersEvent);
 
-        this.client = new Client();
-        this.client.construct();
-        this.client.initializeClient();
+        Client.initialize();
+        Client.initializeClient();
 
         if (ModList.get().isLoaded(Constants.CLOTH_CONFIG_NEOFORGE))
             container.registerExtensionPoint(IConfigScreenFactory.class, new ModConfigMenu());
@@ -38,11 +33,5 @@ public final class NeoForgeMod {
         // Add the overlay manager to the render layers of Gui
         OverlayManager overlayManager = ContainerManager.resolve(OverlayManager.class);
         event.registerBelowAll(Constants.asId("layer/overlaymanager"), overlayManager::render);
-    }
-
-    @SubscribeEvent
-    public void onInitializeClient(FMLClientSetupEvent setupEvent) {
-        // Boot the mod
-        //this.client.initializeClient();
     }
 }